[1946] Refactor into five smaller concerns
Needed to pass the linter
